Added a few missing casts that only show up now for the first time, for some reason.  The lines of code don't appear to have been changed by the recent merges.  Possibly, it is just VS2005 giving a standard warning that I am not used to seeing since I am usually using VS2008.
Also deleted 2 unused local variables that got left over from recently merged fixes.

git-svn-id: https://216.105.38.17/p/tinycad/code/TinyCAD/trunk@487 41062d70-b26d-460a-9302-f438a8930db8

+++ src/Ruler.cpp
@@ -160,12 +160,12 @@ void Ruler::OnPaint()
 	if (dir == 2)
 	{
 		Origin = static_cast<int>(GetTransform().GetOrigin().x);
-		org = m_pDesign->GetCurrentSheet()->GetOptions()->GetOrigin().x;
+		org = (long) m_pDesign->GetCurrentSheet()->GetOptions()->GetOrigin().x;
 	}
 	else
 	{
 		Origin = static_cast<int>(GetTransform().GetOrigin().y);
-		org = m_pDesign->GetCurrentSheet()->GetOptions()->GetOrigin().y;
+		org = (long) m_pDesign->GetCurrentSheet()->GetOptions()->GetOrigin().y;
 	}
 	org *= 10;
